backend-server/auth-service/requirements.txt
flask==3.0.3
flask-cors==4.0.1
ldap3==2.9.1
psycopg2-binary==2.9.9
pyjwt==2.8.0
python-dotenv==1.0.1
